Safer inserts.
