+++ src/test/scala/me/reminisce/database/DatabaseTestHelper.scala
@@ -1,17 +1,8 @@
 package me.reminisce.database
 
-import java.util.concurrent.TimeUnit
-
-import akka.actor.ActorSystem
-import akka.testkit.{ImplicitSender, TestKit}
 import com.github.simplyscala.{MongoEmbedDatabase, MongodProps}
-import com.typesafe.config.ConfigFactory
 import me.reminisce.TestsConfig
-import org.scalatest._
 import reactivemongo.api.{DefaultDB, MongoDriver}
-
-import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, Awaitable}
 
 object DatabaseTestHelper extends MongoEmbedDatabase {
 
@@ -78,41 +69,3 @@ object DatabaseTestHelper extends MongoEmbedDatabase {
   }
 }
 
-abstract class DatabaseTester(actorSystemName: String) extends TestKit(ActorSystem(actorSystemName, ConfigFactory.parseString("akka.loglevel = ERROR")))
-with ImplicitSender
-with WordSpecLike with BeforeAndAfterAll with BeforeAndAfterEach {
-
-  val attemptsPermitted = 20
-
-  import scala.concurrent.ExecutionContext.Implicits.global
-
-  val db = DatabaseTestHelper.getDb
-
-  override def afterAll() {
-    TestKit.shutdownActorSystem(system)
-  }
-
-  override def afterEach(): Unit = {
-    db.drop()
-  }
-
-  def waitAttempts[T](operation: Awaitable[Option[T]], value: Option[T] = None, attempts: Int = 0)
-                     (check: T => Boolean): Option[T] = value match {
-    case None =>
-      if (attempts < attemptsPermitted) {
-        Thread.sleep(200)
-        val newValue = Await.result(operation, Duration(10, TimeUnit.SECONDS))
-        waitAttempts[T](operation, newValue, attempts + 1)(check)
-      } else {
-        None
-      }
-    case Some(result) =>
-      if (check(result)) {
-        Some(result)
-      } else {
-        Thread.sleep(200)
-        val newValue = Await.result(operation, Duration(10, TimeUnit.SECONDS))
-        waitAttempts[T](operation, newValue, attempts + 1)(check)
-      }
-  }
-}
