+++ src/test/scala/me/reminisce/service/gameboardgen/questiongen/OrderByPageLikesSpec.scala
@@ -48,7 +48,7 @@ class OrderByPageLikesSpec extends DatabaseTester("OrderByPageLikesSpec") {
       (0 until pagesNumber) foreach {
         case nb =>
           val selector = BSONDocument("pageId" -> itemIds(nb))
-          Await.result(pagesCollection.update(selector, pages(nb), upsert = true), Duration(10, TimeUnit.SECONDS))
+          Await.result(pagesCollection.save(pages(nb), safeLastError), Duration(10, TimeUnit.SECONDS))
       }
 
       val actorRef = TestActorRef(OrderByPageLikes.props(db))
