+++ src/test/scala/me/reminisce/service/gameboardgen/questiongen/OrderByPostCommentsNumberSpec.scala
@@ -48,7 +48,7 @@ class OrderByPostCommentsNumberSpec extends DatabaseTester("OrderByPostCommentsN
       (0 until postsNumber) foreach {
         case nb =>
           val selector = BSONDocument("userId" -> userId, "postId" -> itemIds(nb))
-          Await.result(pagesCollection.update(selector, posts(nb), upsert = true), Duration(10, TimeUnit.SECONDS))
+          Await.result(pagesCollection.save(posts(nb), safeLastError), Duration(10, TimeUnit.SECONDS))
       }
       val testProbe = TestProbe()
       val actorRef = TestActorRef(OrderByPostCommentsNumber.props(db))
