+++ src/test/scala/me/reminisce/service/gameboardgen/questiongen/WhenDidYouLikeThisPageSpec.scala
@@ -40,7 +40,7 @@ class WhenDidYouLikeThisPageSpec extends DatabaseTester("WhenDidYouLikeThisPageS
       val pageLikesCollection = db[BSONCollection](MongoDatabaseService.fbPageLikesCollection)
       val likedTime = DateTime.now
       val pageLike = FBPageLike(None, userId, itemId, likedTime)
-      Await.result(pageLikesCollection.update(selectorLike, pageLike, upsert = true), Duration(10, TimeUnit.SECONDS))
+      Await.result(pageLikesCollection.save(pageLike, safeLastError), Duration(10, TimeUnit.SECONDS))
 
       val actorRef = TestActorRef(WhenDidYouLikeThisPage.props(db))
       val testProbe = TestProbe()
@@ -55,13 +55,13 @@ class WhenDidYouLikeThisPageSpec extends DatabaseTester("WhenDidYouLikeThisPageS
 
       val selectorPage = BSONDocument("pageId" -> itemId)
       val page = FBPage(None, itemId, Some(s"Cool page with id ID"), None, 1)
-      Await.result(pagesCollection.update(selectorPage, page, upsert = true), Duration(10, TimeUnit.SECONDS))
+      Await.result(pagesCollection.save(page, safeLastError), Duration(10, TimeUnit.SECONDS))
 
       val selectorLike = BSONDocument("userId" -> userId, "pageId" -> itemId)
       val pageLikesCollection = db[BSONCollection](MongoDatabaseService.fbPageLikesCollection)
       val likedTime = DateTime.now
       val pageLike = FBPageLike(None, userId, itemId, likedTime)
-      Await.result(pageLikesCollection.update(selectorLike, pageLike, upsert = true), Duration(10, TimeUnit.SECONDS))
+      Await.result(pageLikesCollection.save(pageLike, safeLastError), Duration(10, TimeUnit.SECONDS))
 
       val actorRef = TestActorRef(WhenDidYouLikeThisPage.props(db))
       val testProbe = TestProbe()
