+++ src/test/scala/me/reminisce/service/gameboardgen/questiongen/WhichCoordinatesWereYouAtSpec.scala
@@ -38,7 +38,7 @@ class WhichCoordinatesWereYouAtSpec extends DatabaseTester("WhichCoordinatesWere
 
       val fbPost = FBPost(postId = itemId, userId = userId)
       val selector = BSONDocument("userId" -> userId, "postId" -> itemId)
-      Await.result(postsCollection.update(selector, fbPost, upsert = true), Duration(10, TimeUnit.SECONDS))
+      Await.result(postsCollection.save(fbPost, safeLastError), Duration(10, TimeUnit.SECONDS))
 
       val actorRef = TestActorRef(WhichCoordinatesWereYouAt.props(db))
       val testProbe = TestProbe()
@@ -59,7 +59,7 @@ class WhichCoordinatesWereYouAtSpec extends DatabaseTester("WhichCoordinatesWere
       val place = FBPlace(None, name = "SuperPlace", location = location, None)
       val selector = BSONDocument("userId" -> userId, "pageId" -> itemId)
       val fbPost = FBPost(postId = itemId, userId = userId, message = Some(postMessage), place = Some(place))
-      Await.result(postsCollection.update(selector, fbPost, upsert = true), Duration(10, TimeUnit.SECONDS))
+      Await.result(postsCollection.save(fbPost, safeLastError), Duration(10, TimeUnit.SECONDS))
 
       val actorRef = TestActorRef(WhichCoordinatesWereYouAt.props(db))
       val testProbe = TestProbe()
